+++ server/index.js
@@ -5,7 +5,8 @@ const cron = require('node-cron')
 const routes = require('@utils/routes')
 const logger = require('@utils/logger')
 const { PORT, inProduction, inDevelopment } = require('@utils/common')
-const { fakeShibbo, requestLogger } = require('./utils/middleware')
+const { fakeShibbo } = require('./utils/fakeshibbo')
+const { requestLogger } = require('./utils/middleware')
 
 const processNewCompletions = require('./scripts/processNewCompletions')
 const processOldCompletions = require('./scripts/processOldCompletions')
@@ -13,6 +14,7 @@ const processOldCompletions = require('./scripts/processOldCompletions')
 const courseCodes = ['AYTKT21018', 'AYTKT21018fi', 'AYTKT21018sv']
 
 const app = express()
+app.use(bodyParser.json({ limit: '5mb' }))
 /**
  * Use hot loading when in development, else serve the static content
  */
@@ -25,6 +27,8 @@ if (inDevelopment) {
   const compiler = webpack(webpackConf('development', { mode: 'development' }))
   app.use(middleware(compiler))
   app.use(hotMiddleWare(compiler))
+  app.use(fakeShibbo)
+  app.use(requestLogger)
 } else {
   app.use('/', express.static('dist/'))
 }
@@ -78,10 +82,6 @@ if (inProduction) {
     checkOodiEntries()
   }) */
 }
-app.use(bodyParser.json({ limit: '5mb' }))
-
-app.use(fakeShibbo)
-app.use(requestLogger)
 
 app.use('/api', routes)
